extended generate_component_names.py to use infiles, should be quite nice to autogenerate methods now. Also it can parse entire components to/from json automatically as long as the files are updated by generate_component_names.py

+++ src/MessageList.cpp
@@ -20,7 +20,7 @@ string Message::print() const
 	stringstream ss;
 	ss << "src: " << src_id
 		<< " dest: " << dest_id
-		<< " type: " << type_text[type]
+		<< " type: " << type_text[static_cast<int>(type)]
 		<< " contents: " << message;
 	return ss.str();
 }
